+++ razorpay/__init__.py
@@ -11,8 +11,12 @@ from .resources import Transfer
 from .resources import VirtualAccount
 from .resources import Addon
 from .resources import Subscription
+from .resources import RegistrationLink
 from .resources import Plan
 from .resources import Settlement
+from .resources import Item
+from .resources import Qrcode
+from .resources import FundAccount
 from .utility import Utility
 from .constants import ERROR_CODE
 from .constants import HTTP_STATUS_CODE
@@ -32,8 +36,12 @@ __all__ = [
         'VirtualAccount',
         'Addon',
         'Subscription',
+        'RegistrationLink',
         'Plan',
+        'FundAccount',
         'Settlement',
+        'Item',
+        'Qrcode',
         'HTTP_STATUS_CODE',
         'ERROR_CODE',
 ]