+++ razorpay/resources/__init__.py
@@ -11,7 +11,11 @@ from .virtual_account import VirtualAccount
 from .addon import Addon
 from .plan import Plan
 from .subscription import Subscription
+from .qrcode import Qrcode
+from .registration_link import RegistrationLink
 from .settlement import Settlement
+from .item import Item
+from .fund_account import FundAccount
 
 __all__ = [
     'Payment',
@@ -27,5 +31,9 @@ __all__ = [
     'Addon',
     'Plan',
     'Subscription',
+    'RegistrationLink',
     'Settlement',
+    'Item',
+    'QrCode',
+    'FundAccount'
 ]