+++ razorpay/resources/virtual_account.py
@@ -1,5 +1,6 @@
 from .base import Resource
 from ..constants.url import URL
+import json
 
 
 class VirtualAccount(Resource):
@@ -69,3 +70,38 @@ class VirtualAccount(Resource):
         """
         url = "{}/{}/payments".format(self.base_url, virtual_account_id)
         return self.get_url(url, data, **kwargs)
+
+    def add_receiver(self, virtual_account_id, data={}, **kwargs):
+        """"
+        Add receiver to an existing virtual account
+
+        Args:
+            virtual_account_id :
+                Id for which Virtual Account objects has to be Closed
+        """
+        url = "{}/{}/receivers".format(self.base_url, virtual_account_id)
+        return self.post_url(url, data, **kwargs)
+
+    def add_allowed_player(self, virtual_account_id, data={}, **kwargs):
+        """"
+        Add an Allowed Payer Account
+
+        Args:
+            virtual_account_id :
+                Id for which Virtual Account objects has to be Closed
+        """
+        url = "{}/{}/allowed_payers".format(self.base_url, virtual_account_id)
+        return self.post_url(url, data, **kwargs) 
+
+    def delete_allowed_player(self, virtual_account_id, allowed_player_id, data={}, **kwargs):
+        """"
+        Delete an Allowed Payer Account
+
+        Args:
+            virtual_account_id :
+                Id for which Virtual Account objects has to be Closed
+        Returns:
+            204        
+        """
+        url = "{}/{}/allowed_payers/{}".format(self.base_url, virtual_account_id, allowed_player_id)
+        return self.delete_url(url, data, **kwargs)